Implement namespacing/auto-loading
