+++ helpers/AltapayHelpers.php
@@ -7,6 +7,8 @@
  * file that was distributed with this source code.
  */
 
+namespace Altapay\Helpers;
+
 require_once dirname( __DIR__, 1 ) . DIRECTORY_SEPARATOR . 'assets' . DIRECTORY_SEPARATOR . 'blade' . DIRECTORY_SEPARATOR . 'lib' . DIRECTORY_SEPARATOR . 'BladeOne.php';
 
 use eftec\bladeone;
@@ -61,7 +63,7 @@ class AltapayHelpers {
 		if ( ! $shippingCountry || ! $shippingCity || ! $shippingAddress || ! $shippingPostal ) {
 			if ( empty( $customerInfo['billing_country'] ) ) {
 				// Throw error since the payment cannot be made without shipping country, even for virtual products
-				return new WP_Error( 'error', __( 'Shipping country is required', 'altapay' ) );
+				return new \WP_Error( 'error', __( 'Shipping country is required', 'altapay' ) );
 			}
 			$customerInfo['shipping_country'] = $customerInfo['billing_country'];
 			$customerInfo['shipping_address'] = $customerInfo['billing_address'];
@@ -81,7 +83,7 @@ class AltapayHelpers {
 	 * @throws Exception
 	 */
 	public function convertDateTimeFormat( $date ) {
-		$dateTime = new DateTime( $date );
+		$dateTime = new \DateTime( $date );
 		return $dateTime->format( 'Y-m-d' );
 	}
